requirements.txt
# MCP framework
fastmcp>=2.0.0

# HTTP
httpx>=0.27.0
requests>=2.32.0

# HTML scraping
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Yahoo Fantasy
yahoo-oauth>=1.0.2
yahoo-fantasy-api>=2.12.2

# ESPN Fantasy (optional)
espn-api>=0.35.0

# Config / secrets
python-dotenv>=1.0.1

# Dev / test
pytest>=8.2.0
pytest-asyncio>=0.23.0
respx>=0.21.0
ruff>=0.4.0
